Uploading half of reactivedrop_latam translation
I'm uploading it to have the achievements and the campaign names for Rich Presence localized on Steam (and also the person who is responsible can now create the missing letters at the end of the mission because a wooden texture appears)
Since I still need to go through the process of proofreading the strings and making sure they work or fit correctly in the game, I have not removed the indents from several translated strings. The only ones from which I have removed the indent are the achievement strings and the campaign names; the rest are indented. I'm not sure if this is the correct approach, please let me know just in case."